+++ src/data/+mag/Instrument.m
@@ -103,7 +103,7 @@ classdef (Sealed) Instrument < handle & matlab.mixin.Copyable & matlab.mixin.Cus
             end
 
             this.cropScience(primaryFilter, secondaryFilter);
-            this.cropDataBasedOnScience();
+            this.cropToMatch();
         end
 
         function cropScience(this, primaryFilter, secondaryFilter)
@@ -120,22 +120,26 @@ classdef (Sealed) Instrument < handle & matlab.mixin.Copyable & matlab.mixin.Cus
             this.Secondary.crop(secondaryFilter);
         end
 
-        function cropDataBasedOnScience(this)
-        % CROPDATABASEDONSCIENCE Crop meta data, events and HK based on
-        % science timestamps.
+        function cropToMatch(this, startTime, endTime)
+        % CROPTOMATCH Crop meta data, events and HK based on science
+        % timestamps or specified timestamps.
 
-            timeRange = this.TimeRange;
+            arguments
+                this (1, 1) mag.Instrument
+                startTime (1, 1) datetime = this.TimeRange(1)
+                endTime (1, 1) datetile = this.TimeRange(2)
+            end
 
             % Filter events.
             if ~isempty(this.Events)
-                this.Events = this.Events(isbetween([this.Events.CommandTimestamp], timeRange(1), timeRange(2), "closed"));
+                this.Events = this.Events(isbetween([this.Events.CommandTimestamp], startTime, endTime, "closed"));
             end
 
             % Filter HK.
-            this.HK.crop(timerange(timeRange(1), timeRange(2), "closed"));
+            this.HK.crop(timerange(startTime, endTime, "closed"));
 
             % Adjust meta data.
-            this.MetaData.Timestamp = timeRange(1);
+            this.MetaData.Timestamp = startTime;
         end
 
         function resample(this, targetFrequency)